+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Update accessibility page following latest accessibility audit.
+
 ## [2.10.0](https://github.com/InseeFr/stromae-dsfr/releases/tag/2.10.0) - 2026-06-26
 
 ### Fixed